Create Complete covid-19 portalREADME.md
Its an webportal which consists of complete covid information like hospitals,availability of beds,ventilators,blood/plasma,oxygen,helpline numbers, location of ambulance,prefered and suggested prescription of doctors,symtoms of covid according to new updates,and clarity about fake news about covid. In this portal we places the new updates and also information regarding the covid. This portal makes people easy to know the information regarding covid.we innovatively thought to make this portal so that people can easily find way to approach for treatment and  be safe.